+++ index.js
@@ -1,9 +1,4 @@
-var crypto = require('crypto')
+var crypto = require('crypto');
 
-var exportFn = require('./pbkdf2')
-var exported = exportFn(crypto)
-
-module.exports = {
-  pbkdf2: exported.pbkdf2,
-  pbkdf2Sync: exported.pbkdf2Sync
-}
+exports.pbkdf2 = crypto.pbkdf2;
+exports.pbkdf2Sync = crypto.pbkdf2Sync;
