+++ package.json
@@ -3,6 +3,7 @@
   "version": "2.0.1",
   "description": "This library provides the functionality of PBKDF2 with the ability to use any supported hashing algorithm returned from crypto.getHashes()",
   "main": "./index.js",
+  "browser": "./browser.js",
   "keywords": [
     "pbkdf2",
     "kdf",
@@ -10,10 +11,10 @@
     "hash"
   ],
   "scripts": {
-    "coverage": "istanbul cover _mocha -- test/*.js",
+    "coverage": "istanbul cover _mocha -- -t 20000 test/index.js",
     "coveralls": "npm run coverage && coveralls < coverage/lcov.info",
-    "test": "istanbul test mocha -- --reporter list test/*.js",
-    "bundle-test": "browserify test/ > test/bundle.js"
+    "test": "mocha --reporter list -t 20000 test/index.js",
+    "bundle-test": "browserify test/index.js > test/bundle.js"
   },
   "repository": {
     "type": "git",
@@ -31,5 +32,7 @@
     "istanbul": "^0.3.5",
     "mocha": "^2.1.0"
   },
-  "dependencies": {}
+  "dependencies": {
+    "create-hmac": "^1.1.0"
+  }
 }
